Adjust bot triggering on reply event
Updated the condition checking statement in the 'MessageCreate' event to correctly check replies to bot messages. Previously, the bot would get triggered even if it wasn't mentioned in the reply as the checking condition was only considering human authors. This commit changes the condition to allow a bot, specifically our bot, to be a valid author of reply messages. This prevents the bot from erroneously triggering when it isn't mentioned and contributes to more efficient bot functionality.