Fix checkGptAvailability parameter issue in MessageCreate
This commit modifies the checkGptAvailability function call within MessageCreate.ts to use the correct user ID. Previously, it was using the ID of the author of the replied message, which resulted in incorrect limit checks for the operation. The pronoun replacement in the returned message was also removed to unify the responses.