Fix incorrect removal of queries from user data
The previous version of code used to incorrectly decrement the number of remaining queries for the users. This resulted in the user's query count becoming less than expected. The correction ensures that the correct number of remaining queries are reflected in the user's data. The fix was implemented by casting the RateLimit variable to a number before performing the decrement operation. This ensures that the subtract operation won't yield an error due to type mismatch.