Add whitelist functionality to GPT service
Implemented whitelist functionality to GPT service in Util.ts. The key-value pair ('userId', true) is now stored when setGptWhitelist is called and checked when checkGptAvailability is called. This allows some specific users to bypass the limit set on the number of GPT-3 queries. The whitelist database is separate from the main 'userData' database, reducing risks of data overwrites.