Refactored Whitelist command to support add, remove and check options
The whitelist command in Whitelist.ts has been expanded to support not only the "add" function but also "remove" and "check". This enhancement facilitates more granular control and transparency over the users who are whitelisted. Now, it's possible to remove a user from the whitelist and check a user's whitelisted status using the same command.