Add deleteGptWhitelist function in Util.ts
Added a new function called deleteGptWhitelist in Utils.ts. This function accepts a user ID and removes the user from the GPT whitelist if they are currently whitelisted. This change will provide support for the 'remove' option in the whitelist command, allowing for more flexibility and control in managing whitelisted users.