Update Queries.ts to limit user access
Updated Queries to restrict non-staff from viewing other users' query data. This was done by checking if the user's role matches any of the roles in the staffRoles from the environment variable. If not a staff member, an error message is returned. This will prevent unauthorized access to sensitive data and ensure user privacy.