Refactor message deletion and add blacklist check
The code for message deletion has been refactored for clarity and maintainability, and a new client parameter was added to the 'messageDelete' method as it was missing previously. Additionally, a check to see if a user is blacklisted has been added. If they are, a message will now be sent informing them of their status and instructing them to contact a moderator, with both messages being deleted after 6000 milliseconds.