Added isAdmin check to control visibility of the Whitelist button
This commit adds a new private variable 'isAdmin' within the Queries class. It is utilized to restrict the visibility of the 'Toggle Whitelist' button only to admin users. This enhancement is to maintain security protocols within the system. Additionally, administrative check logic has been incorporated in the relevant sections with the decision-making process based on the presence of the user's ID in the adminIds variable. Inviting the added 'isAdmin' check led to a more straightforward and securer permission management.