Correct comment typo and remove unnecessary boolean initialization
The typo in the comments for the 'loadAssistant' function was fixed. This is to improve code readability and comprehension. Also, the boolean initialization for the 'existingRun' variable was removed. This was unnecessary since the existence of a run is checked in the subsequent try-catch block and the variable is re-assigned there. Therefore, the initial value does not matter, hence it was removed to make code cleaner.