Add error handling for equal request and response
In both 'Ask.ts' and 'MessageCreate.ts', handling for a case where the response is exactly the same as the query has been added. This is done to avoid sending incorrect responses to the user. If such a situation occurs, an error message will be displayed to the user, informing them to contact the moderation team, along with the error specifics encapsulated within a code block for easy reporting.