Add exclusion of specific channels to bot response conditions
Added a new condition to the `shouldRespond` function in MessageCreate.ts which checks if the bot's message channel is in the list of channels to be excluded from bot responses. This list is retreived from the environment variable 'ExcludedChannels'. This allows for more granular control over where the bot can interact.