Add deferReply at the start of Ask command
This change moves the deferReply function to the start of the Ask command. With this, we're deferring the reply as soon as the function is called, preventing possible issues with late replies due to long processing times. Also, the user is no longer privately notified if there's an ongoing request, making the user interactions smoother and less intrusive.