Add Message object to processGPT function
A new parameter "msg: Message" was added to the "processGPT" function. This change makes it possible for the function to reply directly to the passed message, instead of to the previously hardcoded "message" object. This allows for a more versatile reply handling and hence avoids potential reference errors when trying to reply to different messages.