Refactor message deletion utility function
The deletable check and message delete function have been simplified. Instead of accepting an interaction, delay, and client, it now accepts a message and time. This change was made to improve readability and maintainability of the function by reducing the complexity. Instead of differentiating between messages and command interactions and calling different delete functions accordingly, we now simply check if the message is deletable and delete it if so. The function no longer handles permissions checking, assuming that to be dealt with elsewhere.