Fix incorrect variable reference in MessageCreate event
Changed the variable name from 'message' to 'msg' in the reply method called on an array of responses. The initial variable was causing an error due to reference of a non-existing 'message' variable. This change fixes the bug and ensures the correct 'msg' variable is referenced to reply to messages.